requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile dev.in
#
ansible-compat==24.10.0
    # via
    #   ansible-lint
    #   molecule
ansible-core==2.17.6
    # via
    #   -r ansible-max.in
    #   ansible-compat
    #   ansible-lint
    #   molecule
ansible-lint==24.10.0
    # via -r lint.in
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
black==24.10.0
    # via ansible-lint
bracex==2.5.post1
    # via wcmatch
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.0
    # via tox
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via tox
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   click-help-colors
    #   molecule
    #   pip-tools
click-help-colors==0.9.4
    # via molecule
colorama==0.4.6
    # via tox
cryptography==43.0.3
    # via ansible-core
distlib==0.3.9
    # via virtualenv
distro==1.9.0
    # via selinux
docker==7.1.0
    # via molecule-plugins
enrich==1.2.7
    # via molecule
exceptiongroup==1.2.2
    # via pytest
filelock==3.16.1
    # via
    #   ansible-lint
    #   tox
    #   virtualenv
flake8==7.1.1
    # via -r lint.in
idna==3.10
    # via requests
importlib-metadata==8.5.0
    # via ansible-lint
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   ansible-core
    #   molecule
jsonschema==4.23.0
    # via
    #   ansible-compat
    #   ansible-lint
    #   molecule
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
molecule==24.9.0
    # via
    #   -r molecule.in
    #   molecule-plugins
molecule-plugins[docker]==23.5.3
    # via -r molecule.in
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   ansible-compat
    #   ansible-core
    #   ansible-lint
    #   black
    #   build
    #   molecule
    #   pyproject-api
    #   pytest
    #   tox
pathspec==0.12.1
    # via
    #   ansible-lint
    #   black
    #   yamllint
pip-tools==7.4.1
    # via -r dev.in
platformdirs==4.3.6
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   molecule
    #   pytest
    #   tox
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pyproject-api==1.8.0
    # via tox
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.3
    # via pytest-testinfra
pytest-testinfra==10.1.1
    # via -r molecule.in
pyyaml==6.0.2
    # via
    #   ansible-compat
    #   ansible-core
    #   ansible-lint
    #   molecule
    #   yamllint
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   docker
    #   molecule-plugins
resolvelib==1.0.1
    # via ansible-core
rich==13.9.4
    # via
    #   ansible-lint
    #   enrich
    #   molecule
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via ansible-lint
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
selinux==0.3.0
    # via molecule-plugins
subprocess-tee==0.4.2
    # via
    #   ansible-compat
    #   ansible-lint
tomli==2.1.0
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-api
    #   pytest
    #   tox
tox==4.23.2
    # via -r tox.in
typing-extensions==4.12.2
    # via
    #   black
    #   rich
    #   tox
urllib3==2.2.3
    # via
    #   docker
    #   requests
virtualenv==20.27.1
    # via tox
wcmatch==10.0
    # via
    #   ansible-lint
    #   molecule
wheel==0.45.0
    # via pip-tools
yamllint==1.35.1
    # via
    #   -r lint.in
    #   ansible-lint
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
